Memory.allocate now provides a possibility to force word-aligned allocation.

+++ VM/src/com/sun/max/vm/heap/ImmortalHeap.java
@@ -127,7 +127,7 @@ public final class ImmortalHeap {
      * Initialize the immortal heap memory.
      */
     public static void initialize() {
-        immortalHeap.initialize(maxPermSize.getValue().toInt());
+        immortalHeap.initialize(maxPermSize.getValue());
     }
 
     /**
